Automation Update
Changed the structure to send the number of packets first, then listen until that number of packets are received, then outputs to a new file.

Since we are sending a compressed set of images, adjusted the code to attempt sending a ZIP file instead of a PNG, and included a sample ZIP file for testing.