Transmission changes + MD5
Went back to a simpler transmission method. Image construction ends when a payload is less than 32 bytes. Hopefully this method won't cause problems.

First pass at MD5 sum comparison, sender transmits the MD5 to the receiver, which calculates the MD5 sum of the data it received vs the MD5 sum calculated on the sender side. Prints a success or failure message if it matches, but otherwise no consequences.